+++ R/model_info.R
@@ -709,6 +709,7 @@ model_info.glmmadmb <- function(x, ...) {
 #' @export
 model_info.cpglmm <- function(x, ...) {
   link <- parse(text = .safe_deparse(x@call))[[1]]$link
+  if (is.null(link)) link <- "log"
   if (is.numeric(link)) link <- "tweedie"
   .make_family(
     x = x,
@@ -716,6 +717,22 @@ model_info.cpglmm <- function(x, ...) {
     logit.link = FALSE,
     multi.var = FALSE,
     link.fun = link,
+    ...
+  )
+}
+
+#' @export
+model_info.zcpglm <- function(x, ...) {
+  link <- parse(text = .safe_deparse(x@call))[[1]]$link
+  if (is.null(link)) link <- "log"
+  if (is.numeric(link)) link <- "tweedie"
+  .make_family(
+    x = x,
+    fitfam = "poisson",
+    logit.link = FALSE,
+    multi.var = FALSE,
+    link.fun = link,
+    zero.inf = TRUE,
     ...
   )
 }